Update index.html
💡 Key logic description:
Determination of serial donors: Save the {name, date} object array in localStorage. When the receipt is printed, calculate the number of times the user has won **(current Month)** and automatically stamp [serial donors] if it is more than three times.

Tension Shake: The tension-shake animation applies only to clicked hands, giving you tension while checking.

The title slogan: With each refresh, one of the five slogans is randomly selected to make it fun for the user.