Update index.html
🛠️ Key updates
UI Slimming: Get rid of the complicated preset buttons, and place the GROUP and MEMBERS pop-up buttons left and right.

Strengthen group management: Press the 'GROUP' button to display a list of saved groups, which can be managed through the Delete (×) button next to each group. (Default: Empty)

Filling Intelligent Names: If the name is empty when increasing the headcount, it will be filled with random nicknames automatically.

Minimal Design: The RANDOM, CLEAR and SAVE buttons are compactly placed just above the bottom horizontal line.