Update index.html
Major Modifications
Directed by Central Gathering: Through the shuffle-gather animation, all hands converge and overlap to a point in the center of the screen. At this time, the hands become slightly smaller and tense.

Spread to new position: After a pause in the center, it bounces back to the new position (up, down, left and right) and is in place.

Smooth transition: The cubic-bezier timing function was used to adjust the movement to feel tight and not stiff.